+++ server/src/tangible/__init__.py
@@ -1,3 +1,3 @@
 """Tangible server package."""
 
-__version__ = "0.18.3"
+__version__ = "0.18.4"